Authenticated and Protected Routes Added

+++ client/src/components/ArtworkDetails.js
@@ -1,13 +1,25 @@
 import '../styles/index.css';
 import { useArtworksContext } from '../hooks/useArtworksContext';
+import {useAuthContext} from '../hooks/useAuthContext';
+
+// Date-fns
+import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const ArtworkDetails = ({ artwork, isAdminView }) => {
 
     const { dispatch } = useArtworksContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+        if (!user){
+          return
+        }
+
         const response = await fetch('http://localhost:8080/api/artworks/' + artwork._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+              'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
@@ -21,9 +33,9 @@ const ArtworkDetails = ({ artwork, isAdminView }) => {
         <h4 className='headFont text-xl'>{artwork.title}</h4>
         <span className='font-medium text-sm'>{artwork.desc}</span>
         <span className='headFont c1 text-sm'>{artwork.stat}</span>
-        <span className='headFont c3 text-xs'>{artwork.createdAt}</span>
+        <span className='headFont c3 text-xs'>{formatDistanceToNow(new Date(artwork.createdAt), {addSuffix: true})}</span>
         {isAdminView && (
-          <span className='admin-only-del text-xs text-red-500 bg-gray-300 rounded-lg cursor-pointer text-center w-20 mt-2 hover:bg-red-300' onClick={handleClick}>DELETE</span>
+          <span className='material-symbols-outlined admin-only-del text-xl text-red-500 bg-gray-300 rounded-full cursor-pointer text-center w-8 mt-2 hover:bg-red-300' onClick={handleClick}>DELETE</span>
         )}
       </div>
     )
